protect all back end methods with permissions, fix mappings for interceptor missing, order vacation requests, publishes schedule now, #56

+++ src/main/java/com/sequoiagrove/controller/HolidaysController.java
@@ -1,43 +1,44 @@
 package com.sequoiagrove.controller;
 
 import com.google.gson.*;
+import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.List;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.ui.ModelMap;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.stereotype.Controller;
-import org.springframework.stereotype.Controller;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
-import java.sql.ResultSet;
+import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.sequoiagrove.model.ScheduleTemplate;
-import com.sequoiagrove.model.Day;
-import com.sequoiagrove.model.Scheduled;
 import com.sequoiagrove.controller.MainController;
-
-import com.sequoiagrove.model.RequestStatus;
 import com.sequoiagrove.model.Holiday;
-
-//BAJS_HOLIDAY
-//String
-//hdate
-//name
-//type
 
 @Controller
 public class HolidaysController {
 
+  // extract scope from request
+  @ModelAttribute("scope")
+    public List<String> getId(HttpServletRequest request) {
+      String csvPermissions = (String) request.getAttribute("scope");
+      return Arrays.asList(csvPermissions.split(","));
+    }
+
   @RequestMapping(value = "/holiday")
     public String getAllHolidays(Model model){
+
       JdbcTemplate jdbcTemplate = MainController.getJdbcTemplate();
       List<Holiday> holidayList = jdbcTemplate.query(
           "select * from sequ_holiday",
@@ -58,7 +59,14 @@ public class HolidaysController {
     }
 
   @RequestMapping(value = "/holiday/add")
-    public String sumbitNewHoliday(@RequestBody String data, Model model) throws SQLException {
+    public String sumbitNewHoliday(@RequestBody String data, @ModelAttribute("scope") List<String> permissions, Model model) throws SQLException {
+
+      // the token did not have the required permissions, return 403 status
+      if (!(permissions.contains("manage-employees") || permissions.contains("admin"))) {
+          model.addAttribute("errorStatus", HttpServletResponse.SC_FORBIDDEN);
+          return "jsonTemplate";
+      }
+
       JdbcTemplate jdbcTemplate = MainController.getJdbcTemplate();
       JsonElement jelement = new JsonParser().parse(data);
       JsonObject  jobject = jelement.getAsJsonObject();
@@ -79,7 +87,14 @@ public class HolidaysController {
     }
 
     @RequestMapping(value = "/holiday/update")
-      public String changeRequestDates(@RequestBody String data, Model model) throws SQLException {
+      public String changeRequestDates(@RequestBody String data, @ModelAttribute("scope") List<String> permissions, Model model) throws SQLException {
+
+        // the token did not have the required permissions, return 403 status
+        if (!(permissions.contains("manage-employees") || permissions.contains("admin"))) {
+            model.addAttribute("errorStatus", HttpServletResponse.SC_FORBIDDEN);
+            return "jsonTemplate";
+        }
+
         JdbcTemplate jdbcTemplate = MainController.getJdbcTemplate();
         /*
         Gson gson = new Gson();
@@ -103,18 +118,29 @@ public class HolidaysController {
       }
 
     @RequestMapping(value = "/holiday/remove/{id}")
-      public String deleteRequest(@PathVariable("id") int id, Model model) throws SQLException {
+      public String deleteRequest(@PathVariable("id") int id, @ModelAttribute("scope") List<String> permissions, Model model) throws SQLException {
+        // the token did not have the required permissions, return 403 status
+        if (!(permissions.contains("manage-employees") || permissions.contains("admin"))) {
+            model.addAttribute("errorStatus", HttpServletResponse.SC_FORBIDDEN);
+            return "jsonTemplate";
+        }
         JdbcTemplate jdbcTemplate = MainController.getJdbcTemplate();
         jdbcTemplate.update( "delete from sequ_holiday where id = ?", id);
         return "jsonTemplate";
       }
 
     @RequestMapping(value = "/holiday/get/between/{dateStart}/{dateEnd}")
-      public String getHolidaysBetween(Model model,
+      public String getHolidaysBetween(Model model, @ModelAttribute("scope") List<String> permissions,
           @PathVariable("dateStart") String dateStart,
           @PathVariable("dateEnd") String dateEnd ) throws SQLException{
-        //Make Sure request ID is there too...
 
+        // the token did not have the required permissions, return 403 status
+        if (!(permissions.contains("manage-employees") || permissions.contains("admin"))) {
+            model.addAttribute("errorStatus", HttpServletResponse.SC_FORBIDDEN);
+            return "jsonTemplate";
+        }
+
+        //Make Sure request ID is there too...
         JdbcTemplate jdbcTemplate = MainController.getJdbcTemplate();
         System.out.println(dateStart + " " + dateEnd);
 
@@ -138,7 +164,7 @@ public class HolidaysController {
               return es;
             }
           });
-       
+
         model.addAttribute("holidays", holidayList);
         return "jsonTemplate";
       }
